Part A Bug Fix 2

+++ ImageHoster_assignment_stub/ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -14,12 +14,17 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.PersistenceUnit;
+import javax.persistence.TypedQuery;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.util.*;
 
 @Controller
 public class ImageController {
+
 
     @Autowired
     private ImageService imageService;
@@ -34,6 +39,7 @@ public class ImageController {
         model.addAttribute("images", images);
         return "images";
     }
+
 
     //This method is called when the details of the specific image with corresponding title are to be displayed
     //The logic is to get the image from the databse with corresponding title. After getting the image from the database the details are shown
@@ -92,13 +98,27 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model,HttpSession session) {
         Image image = imageService.getImage(imageId);
+        //Inserting the validation to check if the same user is editing the image
+        Integer image_user =image.getUser().getId();
+        Integer logged_user = ((User) session.getAttribute("loggeduser")).getId();
 
-        String tags = convertTagsToString(image.getTags());
-        model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+        if(!(image_user.equals(logged_user))){
+            String tags = convertTagsToString(image.getTags());
+            String error ="Only owner of the image can edit the image";
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            model.addAttribute("editError",error);
+
+            return "images/image";
+        }
+        else {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -125,9 +145,10 @@ public class ImageController {
             updatedImage.setImageFile(updatedImageData);
         }
 
+
         updatedImage.setId(imageId);
-        User user = (User) session.getAttribute("loggeduser");
-        updatedImage.setUser(user);
+        User logged_user = (User) session.getAttribute("loggeduser");
+        updatedImage.setUser(logged_user);
         updatedImage.setTags(imageTags);
         updatedImage.setDate(new Date());
 
@@ -140,9 +161,30 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId,Model model,HttpSession session) {
+        Image image = imageService.getImage(imageId);
+
+
+        Integer image_user =image.getUser().getId();
+        Integer logged_user = ((User) session.getAttribute("loggeduser")).getId();
+
+        if(!(image_user.equals(logged_user))){
+
+            String tags = convertTagsToString(image.getTags());
+            String error ="Only owner of the image can edit the image";
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            model.addAttribute("deleteError",error);
+            return "images/image";
+        }
+        else {
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", tags);
+            imageService.deleteImage(imageId);
+            return "images";
+        }
+
     }
